+++ lib/src/widgets/rive_animation.dart
@@ -76,6 +76,14 @@ class RiveAnimation extends StatefulWidget {
   /// Default `1.0`.
   final double speedMultiplier;
 
+  /// For Rive Listeners, allows scrolling behavior to still occur on Rive
+  /// widgets when a touch/drag action is performed on touch-enabled devices.
+  /// Otherwise, scroll behavior may be prevented on touch/drag actions on the
+  /// widget by default.
+  ///
+  /// Default `false`.
+  final bool isTouchScrollEnabled;
+
   /// Creates a new [RiveAnimation] from an asset bundle.
   ///
   /// *Example:*
@@ -98,6 +106,7 @@ class RiveAnimation extends StatefulWidget {
     this.behavior = RiveHitTestBehavior.opaque,
     this.objectGenerator,
     this.speedMultiplier = 1,
+    this.isTouchScrollEnabled = false,
     Key? key,
   })  : name = asset,
         file = null,
@@ -128,6 +137,7 @@ class RiveAnimation extends StatefulWidget {
     this.behavior = RiveHitTestBehavior.opaque,
     this.objectGenerator,
     this.speedMultiplier = 1,
+    this.isTouchScrollEnabled = false,
     Key? key,
   })  : name = url,
         file = null,
@@ -156,6 +166,7 @@ class RiveAnimation extends StatefulWidget {
     this.behavior = RiveHitTestBehavior.opaque,
     this.objectGenerator,
     this.speedMultiplier = 1,
+    this.isTouchScrollEnabled = false,
     Key? key,
   })  : name = path,
         file = null,
@@ -185,6 +196,7 @@ class RiveAnimation extends StatefulWidget {
     this.controllers = const [],
     this.onInit,
     this.speedMultiplier = 1,
+    this.isTouchScrollEnabled = false,
     Key? key,
     this.behavior = RiveHitTestBehavior.opaque,
   })  : name = null,
@@ -356,6 +368,7 @@ class RiveAnimationState extends State<RiveAnimation> {
           enablePointerEvents: _shouldAddHitTesting,
           behavior: widget.behavior,
           speedMultiplier: widget.speedMultiplier,
+          isTouchScrollEnabled: widget.isTouchScrollEnabled,
         )
       : widget.placeHolder ?? const SizedBox();
 }